+++ packages/docs/src/components/Navbar/NavbarContent/styles.module.css
@@ -84,7 +84,7 @@
   opacity: 1;
 }
 
-.download p {
+/* .download p {
   font-size: 16px;
   font-weight: 400;
 
@@ -105,7 +105,7 @@
 
 .download:hover {
   opacity: 0.75;
-}
+} */
 
 a:hover {
   text-decoration: none;
@@ -142,43 +142,4 @@ a:hover {
     width: 140px;
     height: 36px;
   }
-}
-dialog.modalContainer {
-  display: flex;
-  flex-direction: column;
-  max-width: 480px;
-  width: 100%;
-  gap: 24px;
-  padding: 24px;
-  background-color: var(--swm-background);
-}
-
-dialog.modalContainer[open]::backdrop {
-  backdrop-filter: blur(5px);
-}
-.modalContainer p {
-  margin: 0;
-}
-.modalHead p {
-  font-size: 24px;
-}
-.modalHead {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex: 1 1 0;
-}
-
-.modalSubheading {
-  color: var(--radon-subheading);
-  font-size: 18px;
-}
-
-.dialogCloseButton {
-  cursor: pointer;
-  background-color: transparent;
-  border: none;
-  display: flex;
-  justify-content: center;
-  align-items: center;
 }
