+++ packages/docs/src/components/Navbar/NavbarDownloadButton/index.tsx
@@ -1,0 +1,16 @@
+import React, { RefObject } from "react";
+import styles from "./styles.module.css";
+
+interface NavbarDownloadButtonProps {
+  dialogRef: RefObject<HTMLDialogElement>;
+}
+export default function NavbarDownloadButton({ dialogRef }: NavbarDownloadButtonProps) {
+  const handleDialogOpen = () => {
+    dialogRef.current?.showModal();
+  };
+  return (
+    <button className={styles.download} onClick={handleDialogOpen}>
+      <p>Download</p>
+    </button>
+  );
+}
